change some datatype and some field name, fix conflict

+++ src/main/java/org/petspa/petcaresystem/authenuser/controller/AuthenUserController.java
@@ -14,7 +14,7 @@ public class AuthenUserController {
     AuthenUserService authenUserService;
 
     @PutMapping("/profile/{id}")
-    public ResponseEntity<ResponseObj> UpdateProflie(@PathVariable int id,
+    public ResponseEntity<ResponseObj> UpdateProflie(@PathVariable Long id,
                                                      @RequestBody UpdateProfileRequest profileRequest){
         return authenUserService.UpdateProflie(id, profileRequest);
     }
